api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
paho-mqtt==2.1.0
pycryptodome==3.23.0
